Move many property methods from Communicator into Properties extension class
